requirements.txt
pyTelegramBotAPI==4.15.4
openai-whisper
ffmpeg-python